feat: theme exports common config
