+++ apify-docs-theme/package.json
@@ -15,6 +15,7 @@
   "license": "ISC",
   "dependencies": {
     "@docusaurus/theme-common": "^2.2.0",
-    "babel-loader": "^9.1.0"
+    "babel-loader": "^9.1.0",
+    "prism-react-renderer": "^1.3.5"
   }
 }
