+++ apify-docs-theme/src/index.js
@@ -1,41 +1,7 @@
-function theme(
-    context,
-    options,
-) {
-    return {
-        name: '@apify/docs-theme',
-        getThemePath() {
-            return '../src/theme';
-        },
-        getTypeScriptThemePath() {
-            return '../src/theme';
-        },
-        contentLoaded({ actions }) {
-            const { setGlobalData } = actions;
-            setGlobalData({
-                options,
-            });
-        },
-        getClientModules() {
-            return [
-                require.resolve('./theme/custom.css'),
-            ];
-        },
-        configureWebpack() {
-            return {
-                module: {
-                    rules: [
-                        {
-                            test: /(@apify\/|apify-)docs-theme\/src\/theme\/.*?\.jsx?$/,
-                            use: {
-                                loader: 'babel-loader',
-                            },
-                        },
-                    ],
-                },
-            };
-        },
-    };
-}
+const config = require('./config.js');
+const { theme } = require('./theme.js');
 
-exports.default = theme;
+module.exports = {
+    default: theme,
+    config,
+};
