+++ apify-docs-theme/src/theme.js
@@ -1,0 +1,43 @@
+function theme(
+    context,
+    options,
+) {
+    return {
+        name: '@apify/docs-theme',
+        getThemePath() {
+            return '../src/theme';
+        },
+        getTypeScriptThemePath() {
+            return '../src/theme';
+        },
+        contentLoaded({ actions }) {
+            const { setGlobalData } = actions;
+            setGlobalData({
+                options,
+            });
+        },
+        getClientModules() {
+            return [
+                require.resolve('./theme/custom.css'),
+            ];
+        },
+        configureWebpack() {
+            return {
+                module: {
+                    rules: [
+                        {
+                            test: /(@apify\/|apify-)docs-theme\/src\/theme\/.*?\.jsx?$/,
+                            use: {
+                                loader: 'babel-loader',
+                            },
+                        },
+                    ],
+                },
+            };
+        },
+    };
+}
+
+module.exports = {
+    theme,
+};
